requirements.txt
langchain>=0.3.0
langchain-community>=0.3.0
langchain-chroma>=0.2.0
langgraph>=0.4.0
chromadb>=0.5.0
langchain-text-splitters>=0.3.0
gradio>=5.0.0
zhipuai>=2.0.0
python-dotenv>=1.0.0
Pillow>=10.0.0
